Renamed cross.py to arithmetic.py per suggestion by @paulromano
